requirements.txt
setuptools
greenlet
bottle
gevent
gevent-websocket
RPi.GPIO
Adafruit-MAX31855
Adafruit-GPIO
websocket-client
paho.mqtt
